# Add zstd decompression support