## Changes Made

### zstd decompression

Added support for zstd decompression, indicated by "Content-Encoding: zstd" http request header.

### New benchmark

Added a new benchmark that performs decoding of various batch sizes using all 3  supported Content-Encodings. Benchmark results show zstd uses less CPU than GZIP for all batch sizes:

```
name                            time/op
DecodeMatrix/none/batch=1-8     3.35µs ± 3%
DecodeMatrix/gzip/batch=1-8     6.26µs ± 1%
DecodeMatrix/zstd/batch=1-8     4.45µs ± 4%
DecodeMatrix/none/batch=10-8    12.2µs ± 0%
DecodeMatrix/gzip/batch=10-8    23.4µs ± 3%
DecodeMatrix/zstd/batch=10-8    15.5µs ± 5%
DecodeMatrix/none/batch=100-8    100µs ±10%
DecodeMatrix/gzip/batch=100-8    141µs ±10%
DecodeMatrix/zstd/batch=100-8    123µs ± 4%
DecodeMatrix/none/batch=1000-8   964µs ± 5%
DecodeMatrix/gzip/batch=1000-8  1.30ms ± 4%
DecodeMatrix/zstd/batch=1000-8  1.00ms ± 4%

name                            alloc/op
DecodeMatrix/none/batch=1-8     6.84kB ± 0%
DecodeMatrix/gzip/batch=1-8     11.1kB ± 0%
DecodeMatrix/zstd/batch=1-8     6.87kB ± 0%
DecodeMatrix/none/batch=10-8    18.8kB ± 0%
DecodeMatrix/gzip/batch=10-8    23.2kB ± 0%
DecodeMatrix/zstd/batch=10-8    18.9kB ± 0%
DecodeMatrix/none/batch=100-8    138kB ± 0%
DecodeMatrix/gzip/batch=100-8    144kB ± 0%
DecodeMatrix/zstd/batch=100-8    139kB ± 0%
DecodeMatrix/none/batch=1000-8  1.34MB ± 0%
DecodeMatrix/gzip/batch=1000-8  1.35MB ± 0%
DecodeMatrix/zstd/batch=1000-8  1.39MB ± 1%

name                            allocs/op
DecodeMatrix/none/batch=1-8       36.0 ± 0%
DecodeMatrix/gzip/batch=1-8       39.0 ± 0%
DecodeMatrix/zstd/batch=1-8       37.0 ± 0%
DecodeMatrix/none/batch=10-8       157 ± 0%
DecodeMatrix/gzip/batch=10-8       160 ± 0%
DecodeMatrix/zstd/batch=10-8       158 ± 0%
DecodeMatrix/none/batch=100-8    1.33k ± 0%
DecodeMatrix/gzip/batch=100-8    1.37k ± 0%
DecodeMatrix/zstd/batch=100-8    1.33k ± 0%
DecodeMatrix/none/batch=1000-8   13.0k ± 0%
DecodeMatrix/gzip/batch=1000-8   13.1k ± 0%
DecodeMatrix/zstd/batch=1000-8   13.0k ± 0%
```

### Eliminated tmp buffer

Removed `tmp` field from pool object and eliminated io.CopyBuffer. This is not necessary since io.CopyBuffer does not use a temp buffer if either dest or src are capable of ReadFrom/WriteTo shortcut (see https://cs.opensource.google/go/go/+/master:src/io/io.go;drc=1ff89009f198ad5bae3549dd3b992882bd97e5f8;l=411)

We now use io.Copy instead. Benchmarking before and after the change shows that there is no performance change. I kept the existing BenchmarkDecode function to make sure we can easily compare the performance before and after.

Performance Comparison:

```
name      old time/op    new time/op    delta
Decode-8    72.5µs ± 2%    71.3µs ± 2%    ~     (p=0.210 n=10+10)

name      old alloc/op   new alloc/op   delta
Decode-8    83.3kB ± 0%    83.0kB ± 0%  -0.46%  (p=0.000 n=10+10)

name      old allocs/op  new allocs/op  delta
Decode-8     1.08k ± 0%     1.08k ± 0%    ~     (all equal)
```